server/requirements.txt
blinker==1.9.0
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
packaging==24.2
pillow==11.0.0
python-dotenv==1.0.1
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
Werkzeug==3.1.3
